pyproject.toml
[project]
name = "drf-course-api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "qe-present", email = "2664481691@qq.com" }
]
dependencies = [
    "djangorestframework>=3.15.2",
    "pillow>=11.1.0",
    "django-extensions>=3.2.3",
    "django-silk>=5.4.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "django-filter>=25.1",
    "isort>=6.0.1",
    "black>=25.1.0",
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []
[tool.rye.scripts]
dev = "python manage.py runserver"
auto="gitsync/gitupd.bat"

[tool.hatch.metadata]
allow-direct-references = true


[tool.hatch.build.targets.wheel]
packages = ["src/drf_course_api"]
